Add macOS `arm64` support (#5)
This PR adds support for building v8go for the `arm64` architecture on macOS.

It should be straightforward to add support for Linux `arm64` with the same approach, however I'm punting on this for now.  May be able to run this in GHA since V8 supports cross-compiling (`target_cpu` vs `v8_target_cpu`). May actually be able to do the same for macOS, but we would only be able to do the build (not the tests).

In a future PR, I'll hook up a self-hosted runner on our M1 mini to perform the "V8 Build" GHA for macOS arm64.